cleanup `particlesFunctors.hpp`
To reduce the compile time for independent compile units solver calls were moved to the corresponding stages.
