+++ include/picongpu/particles/ionization/byField/IonizationCurrent/JIonizationAssignment.hpp
@@ -22,7 +22,6 @@
 #include "picongpu/simulation_defines.hpp"
 
 #include "picongpu/fields/FieldJ.hpp"
-#include "picongpu/particles/ParticlesFunctors.hpp"
 
 namespace picongpu
 {
